perf: Narrow memory usage and started porting the data to State

+++ Drakhtar/Structures/UnitFactory.cpp
@@ -59,7 +59,7 @@ UnitFactory::UnitFactory(Scene* scene) : scene_(scene) {
 UnitFactory::~UnitFactory() = default;
 
 Unit* UnitFactory::newBattalion(const std::string& type, Team* team, Box* box,
-                                int size) {
+                                byte size) {
   if (statMap.find(type) == statMap.end()) return nullptr;
   const auto textureName =
       (team->getColor() == Color::BLUE ? "Units-Blue" : "Units-Red") + type;
